lab32/blogapi-permissions/pyproject.toml
[tool.poetry]
name = "blogapi-permissions"
version = "0.1.0"
description = ""
authors = ["Eman Al-Shaikh Hussain <engemanalsheikh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
